requirements.txt
# Core framework
crewai[anthropic]>=1.9.0

# Data processing (Python 3.13 compatible versions)
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Database & state
databricks-sql-connector>=3.0.0
redis>=5.0.0

# Configuration & validation
pydantic>=2.6.1
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# LLM clients
openai>=1.10.0
anthropic>=0.18.0
sentence-transformers>=2.3.1

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=8.0.0

# Unification layer (install with: pip install -e ../ergonosis_unification)
ergonosis-unification
